Added multi tenancy support
General:
Adapted function to forward the tenant id to ddb, s3 and es
DynamoDb:
This approach allows multitenancy by using only one dynamodb table.
the Primary Key is a combined id, which is a concatenation of resource id and tenant id.
In order to restore the resource id again the tenant id is also stored separately.
Elastic search:
The ddbToEs takes the restored resource id and the tenant id for indexing.
s3 object storage:
The filename will also include the tenant id.

Test:
Added / refactored tests in order to cover the multi tenant usecase